Add Router API Endpoints Modules
