+++ demo/main2.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" type="text/css" href="main.css">
     <title>Welcome to Filip Filchev's Site</title>
-    <script src="http://localhost:8080/sdk"></script>
     <style>/* Global Styles */
     body {
         background-color: #5D3FD3;
@@ -190,6 +189,7 @@
 <footer class="footer">
     <p>&copy; 2023 Filip Filchev's Site. All rights reserved.</p>
 </footer>
+<script src="http://localhost:8080/api/sdk"></script>
 <script>
   sdk.setChatBoxDivId('chat-container')
   sdk.createLoginForm();
